Footer and Navbar: moves files into src/components/layouts; index: updates Footer and Navbar import path
